+++ src/core/geometry.rs
@@ -1,21 +1,22 @@
+use crate::{List, Vector, Vectors};
 use itertools::multizip;
-use na::{convert, Matrix1xX, Matrix3x1, Matrix3xX, RealField};
+use na::{convert, RealField};
 
-/// Magnitudes of a list of vectors 3xX.
-pub fn magnitudes<T>(vectors: &Matrix3xX<T>) -> Matrix1xX<T>
+/// Magnitudes of a list of vectors.
+pub fn magnitudes<T>(vectors: &Vectors<T>) -> List<T>
 where
     T: RealField,
 {
-    let number_vectors = vectors.ncols();
-    let mut magnitudes = Matrix1xX::<T>::zeros(number_vectors);
-    for (res, vector) in multizip((magnitudes.iter_mut(), vectors.column_iter())) {
+    let size = crate::number_vectors(vectors);
+    let mut magnitudes = List::<T>::zeros(size);
+    for (res, vector) in multizip((magnitudes.iter_mut(), vectors.row_iter())) {
         *res = vector.norm();
     }
     magnitudes
 }
 
-/// Distances from of a list of vectors 3xX to another one.
-pub fn distances<T>(vectors_1: &Matrix3xX<T>, vectors_2: &Matrix3xX<T>) -> Matrix1xX<T>
+/// Distances from of a list of vectors to another one.
+pub fn distances<T>(vectors_1: &Vectors<T>, vectors_2: &Vectors<T>) -> List<T>
 where
     T: RealField,
 {
@@ -23,7 +24,7 @@ where
 }
 
 /// Distance from a vector 3x1 to another one.
-pub fn distance<T>(vector_1: &Matrix3x1<T>, vector_2: &Matrix3x1<T>) -> T
+pub fn distance<T>(vector_1: &Vector<T>, vector_2: &Vector<T>) -> T
 where
     T: RealField,
 {
@@ -31,20 +32,20 @@ where
 }
 
 /// Unit vectors of a list of vectors 3xX.
-pub fn units<T>(vectors: &Matrix3xX<T>) -> Matrix3xX<T>
+pub fn units<T>(vectors: &Vectors<T>) -> Vectors<T>
 where
     T: RealField,
 {
-    let number_vectors = vectors.ncols();
-    let mut directions = Matrix3xX::zeros(number_vectors);
-    for (mut direction, vector) in multizip((directions.column_iter_mut(), vectors.column_iter())) {
+    let size = crate::number_vectors(vectors);
+    let mut directions = Vectors::zeros(size);
+    for (mut direction, vector) in multizip((directions.row_iter_mut(), vectors.row_iter())) {
         direction.copy_from(&vector.normalize());
     }
     directions
 }
 
 /// Directions from of a list of vectors 3xX to another one.
-pub fn directions<T>(vectors_1: &Matrix3xX<T>, vectors_2: &Matrix3xX<T>) -> Matrix3xX<T>
+pub fn directions<T>(vectors_1: &Vectors<T>, vectors_2: &Vectors<T>) -> Vectors<T>
 where
     T: RealField,
 {
@@ -52,7 +53,7 @@ where
 }
 
 /// Direction from a vector 3x1 to another one.
-pub fn direction<T>(vector_1: &Matrix3x1<T>, vector_2: &Matrix3x1<T>) -> Matrix3x1<T>
+pub fn direction<T>(vector_1: &Vector<T>, vector_2: &Vector<T>) -> Vector<T>
 where
     T: RealField,
 {
@@ -60,15 +61,15 @@ where
 }
 
 /// Component-wise cartesian to spherical conversion.
-pub fn cart_to_sph<T>(matrix: &Matrix3xX<T>) -> Matrix3xX<T>
+pub fn cart_to_sph<T>(vectors: &Vectors<T>) -> Vectors<T>
 where
     T: RealField,
 {
-    let mut sphericals = Matrix3xX::zeros(matrix.ncols());
-    for (mut spherical, cartesian) in multizip((sphericals.column_iter_mut(), matrix.column_iter()))
-    {
+    let size = crate::number_vectors(vectors);
+    let mut sphericals = Vectors::zeros(size);
+    for (mut spherical, cartesian) in multizip((sphericals.row_iter_mut(), vectors.row_iter())) {
         if cartesian.norm() == convert(0.) {
-            spherical.copy_from(&Matrix3x1::<T>::zeros());
+            spherical.copy_from(&Vector::<T>::zeros());
         } else {
             spherical.copy_from_slice(&[
                 cartesian[1].atan2(cartesian[0]),